Create PR with version change to main (#11)
This PR introduces an extension to the release Github workflow.
The workflow will now introduce the Maven version bump in a separate
branch and open a PR against main, in order to advertise the version change
in the source code. 
Since the main branch is usually protected, this is the only way.